requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
Flask-WTF==0.14.3
python-dateutil==2.8.1
babel==2.9.0
WTForms==2.3.3
alembic==1.4.3
SQLAlchemy==1.3.20
Werkzeug==1.0.1
Jinja2==2.11.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0